CHANGEND: Type of weight from long long to float.

+++ src/graph.c
@@ -1,12 +1,12 @@
 #include "graph.h"
 
 
-Graph createGraph(unsigned long numberOfVertices, unsigned long numberOfEdges) {
+Graph createGraph(long numberOfVertices, long numberOfEdges) {
     Graph graph = {};
     graph.numberOfEdges = numberOfEdges;
     graph.numberOfVertices = numberOfVertices;
     graph.edgeListSize = numberOfEdges * 3;
-    graph.edgeList = (long long *) malloc(sizeof(long long) * graph.edgeListSize);
+    graph.edgeList = (float *) malloc(sizeof(float) * graph.edgeListSize);
     for (unsigned long i = 0; i < graph.edgeListSize; i++) {
         graph.edgeList[i] = EDGE_NOT_INIT;
     }
@@ -15,9 +15,9 @@ Graph createGraph(unsigned long numberOfVertices, unsigned long numberOfEdges) {
     return graph;
 }
 
-void addEdge(Graph *graph, unsigned long vertex1, unsigned long vertex2, long long weight) {
-    graph->edgeList[graph->edgePointer] = (long long) vertex1;
-    graph->edgeList[graph->edgePointer+1] = (long long) vertex2;
+void addEdge(Graph *graph, long vertex1, long vertex2, float weight) {
+    graph->edgeList[graph->edgePointer] = vertex1;
+    graph->edgeList[graph->edgePointer+1] = vertex2;
     graph->edgeList[graph->edgePointer+2] = weight;
     graph->edgePointer += 3;
